slice viewStore with update and utility slices
